Create playlistApp
Move actual player to another place.
Renamed the remaining of the Player component to PlaylistApp.
