+++ src/jsx/components/playlistApp/player/ManageButton.jsx
