+++ src/jsx/components/playlistApp/playlist/Entry.jsx
